+++ src/components/index.js
@@ -2,4 +2,6 @@
 // IDEA: Load components dynamically via Weback's req.context?
 
 export { default as App } from './App/App';
+export { default as Board } from './Board/Board';
 export { default as Message } from './Message/Message';
+export { default as Thread } from './Thread/Thread';